third_party/WebKit/LayoutTests/inspector/styles/stylesheet-tracking-expected.txt

Tests that the styleSheetAdded and styleSheetRemoved events are reported into the frontend. Bug 105828.

Text
2 headers known:
    URL=.../inspector/styles/resources/stylesheet-tracking.css
    origin=regular
    URL=.../inspector/styles/stylesheet-tracking.html
    origin=regular
=== Adding stylesheet... ===
Stylesheets added:
    URL=.../inspector/styles/resources/styles-new-API-1.css
    origin=regular
    URL=.../inspector/styles/resources/styles-new-API-2.css
    origin=regular
    URL=.../inspector/styles/resources/styles-new-API.css
    origin=regular
    URL=.../inspector/styles/stylesheet-tracking.html
    origin=regular
=== Removing @import... ===
Stylesheets removed:
    URL=.../inspector/styles/resources/styles-new-API-1.css
    origin=regular
    URL=.../inspector/styles/resources/styles-new-API-2.css
    origin=regular
    URL=.../inspector/styles/resources/styles-new-API.css
    origin=regular
=== Removing stylesheet... ===
Stylesheets removed:
    URL=.../inspector/styles/stylesheet-tracking.html
    origin=regular
=== Adding rule... ===
Stylesheets added:
    URL=.../inspector/styles/stylesheet-tracking.html
    origin=inspector
Rule added
